backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pandas>=2.2.0
numpy>=2.0.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-multipart>=0.0.9
aiofiles>=23.2.1
gunicorn>=22.0.0
python-dotenv>=1.0.0
haversine>=2.8.1
